+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# <a href='https://github.com/mrjackwills/belugasnooze_pi/releases/tag/v0.1.3'>v0.1.3</a>
+### 2022-10-08
+
+### Chores
++ dev container updated, [81b917a9](https://github.com/mrjackwills/belugasnooze_pi/commit/81b917a9c4029cda3a3cbb2c59cd3f39f3c92d70),
+
+### Refactors
++ dead code removed, [24c0578d](https://github.com/mrjackwills/belugasnooze_pi/commit/24c0578db006b6e6684a10a5bd236c44da0d0535),
++ into_iter().map, [522185e6](https://github.com/mrjackwills/belugasnooze_pi/commit/522185e65844caefac9c53ae86e4873e455a3dc7),
++ remove async from parse_env, [66f7f9aa](https://github.com/mrjackwills/belugasnooze_pi/commit/66f7f9aaa0128cc164d873253f06bf025eb6b2e8),
+
+
 # <a href='https://github.com/mrjackwills/belugasnooze_pi/releases/tag/v0.1.2'>v0.1.2</a>
 ### 2022-09-27